feat: 1.combine the LoadOption function; 2.support load with default if necessary.
